requirements.txt
# Core
gitpython==3.1.40
tree-sitter==0.20.1
pygments==2.17.2

# Vector/Embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2
chromadb==0.4.24

# LLM
transformers==4.38.2
accelerate==0.27.2
bitsandbytes==0.43.0

# CLI
click==8.1.7
rich==13.7.0